🚚 rename: userTownContext -> userLocationContext 네이밍 수정
